requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
xgboost==3.0.4
lightgbm==4.6.0
mlflow==3.10.0
matplotlib==3.9.2
streamlit==1.37.1
shap==0.48.0
PyYAML==6.0.1
joblib==1.4.2
pyarrow>=16.1.0
tabulate>=0.9.0
pytest==7.4.4
